+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 * Fixed issue with updating navigation after editing an output.
 * Added option to expand criteria forms on search forms.
 * Added a Date field.
+* Added function to clone a data processor.
 
 # Version 1.0.7